scrubbed "Bing" from _inside_ all files, renamed package(s)
