+++ Sample/Android/Xamarin.Cognitive.Speech.Sample.Android.csproj
@@ -7,8 +7,8 @@
     <ProjectTypeGuids>{EFBA0AD7-5A72-4C68-AF49-83D382785DCF};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <TemplateGuid>{c9e5eea5-ca05-42a1-839b-61506e0a37df}</TemplateGuid>
     <OutputType>Library</OutputType>
-    <RootNamespace>Xamarin.Cognitive.BingSpeech.Sample.Droid</RootNamespace>
-    <AssemblyName>Xamarin.Cognitive.BingSpeech.Sample.Android</AssemblyName>
+    <RootNamespace>Xamarin.Cognitive.Speech.Sample.Droid</RootNamespace>
+    <AssemblyName>Xamarin.Cognitive.Speech.Sample.Android</AssemblyName>
     <AndroidApplication>True</AndroidApplication>
     <AndroidResgenFile>Resources\Resource.designer.cs</AndroidResgenFile>
     <AndroidResgenClass>Resource</AndroidResgenClass>
@@ -113,7 +113,7 @@
     <Folder Include="Resources\drawable\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Xamarin.Cognitive.BingSpeech.Sample\Xamarin.Cognitive.Speech.Sample.csproj">
+    <ProjectReference Include="..\Xamarin.Cognitive.Speech.Sample\Xamarin.Cognitive.Speech.Sample.csproj">
       <Project>{65c2b7b3-a7ca-4581-9c08-d05f5de61e13}</Project>
       <Name>Xamarin.Cognitive.Speech.Sample</Name>
     </ProjectReference>
