+++ Sample/UWP/App.xaml.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -15,7 +15,7 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-namespace Xamarin.Cognitive.BingSpeech.Sample.UWP
+namespace Xamarin.Cognitive.Speech.Sample.UWP
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
