+++ Sample/UWP/MainPage.xaml
@@ -1,9 +1,9 @@
-﻿<forms:WindowsPage
-    x:Class="Xamarin.Cognitive.BingSpeech.Sample.UWP.MainPage"
+<forms:WindowsPage
+    x:Class="Xamarin.Cognitive.Speech.Sample.UWP.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:forms="using:Xamarin.Forms.Platform.UWP"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Xamarin.Cognitive.BingSpeech.Sample.UWP"
+    xmlns:local="using:Xamarin.Cognitive.Speech.Sample.UWP"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
