+++ Sample/UWP/Package.appxmanifest
@@ -3,7 +3,7 @@
   <Identity Name="2eb876a5-5766-436f-a4c1-181e43cf4bb9" Publisher="CN=20d51863-269a-4c23-8eea-d6c657c294ee" Version="1.0.0.0" />
   <mp:PhoneIdentity PhoneProductId="ec0cc741-fd3e-485c-81be-68815c480690" PhonePublisherId="00000000-0000-0000-0000-000000000000" />
   <Properties>
-    <DisplayName>Xamarin.Cognitive.BingSpeech.Sample.UWP</DisplayName>
+    <DisplayName>Xamarin.Cognitive.Speech.Sample.UWP</DisplayName>
     <PublisherDisplayName>20d51863-269a-4c23-8eea-d6c657c294ee</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -14,8 +14,8 @@
     <Resource Language="x-generate" />
   </Resources>
   <Applications>
-    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="Xamarin.Cognitive.BingSpeech.Sample.UWP.App">
-      <uap:VisualElements DisplayName="Xamarin.Cognitive.BingSpeech.Sample.UWP" Square150x150Logo="Assets\Square150x150Logo.png" Square44x44Logo="Assets\Square44x44Logo.png" Description="Xamarin.Cognitive.BingSpeech.Sample.UWP" BackgroundColor="transparent">
+    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="Xamarin.Cognitive.Speech.Sample.UWP.App">
+      <uap:VisualElements DisplayName="Xamarin.Cognitive.Speech.Sample.UWP" Square150x150Logo="Assets\Square150x150Logo.png" Square44x44Logo="Assets\Square44x44Logo.png" Description="Xamarin.Cognitive.Speech.Sample.UWP" BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png" Square310x310Logo="Assets\LargeTile.png" Square71x71Logo="Assets\SmallTile.png">
           <uap:ShowNameOnTiles>
             <uap:ShowOn Tile="square150x150Logo" />
