--- Sample/Xamarin.Cognitive.BingSpeech.Sample/Keys.cs
@@ -1,12 +1,0 @@
-namespace Xamarin.Cognitive.BingSpeech.Sample
-{
-	public static class Keys
-	{
-		public static class BingSpeech
-		{
-			public const string BadSubscriptionKey = "!BAD KEY!"; // get a key here: https://azure.microsoft.com/en-us/try/cognitive-services/my-apis/?apiSlug=speech-api
-
-			public const string SubscriptionKey = "BadSubscriptionKey"; // replace with your own key here, e.g. "8a29574dc98a4a29566fa7f6440d3d39";
-		}
-	}
-}
