+++ Sample/iOS/AppDelegate.cs
@@ -1,11 +1,11 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
 using Foundation;
 using UIKit;
 
-namespace Xamarin.Cognitive.BingSpeech.Sample.iOS
+namespace Xamarin.Cognitive.Speech.Sample.iOS
 {
     // The UIApplicationDelegate for the application. This class is responsible for launching the 
     // User Interface of the application, as well as listening (and optionally responding) to
