+++ Sample/iOS/Info.plist
@@ -23,15 +23,15 @@
 	<key>MinimumOSVersion</key>
 	<string>8.0</string>
 	<key>CFBundleDisplayName</key>
-	<string>Xamarin.Cognitive.BingSpeech.Sample</string>
+	<string>Xamarin.Cognitive.Speech.Sample</string>
 	<key>CFBundleIdentifier</key>
-	<string>com.companyname.Xamarin.Cognitive.BingSpeech.Sample</string>
+	<string>com.companyname.Xamarin.Cognitive.Speech.Sample</string>
 	<key>CFBundleVersion</key>
 	<string>1.0</string>
 	<key>UILaunchStoryboardName</key>
 	<string>LaunchScreen</string>
 	<key>CFBundleName</key>
-	<string>Xamarin.Cognitive.BingSpeech.Sample</string>
+	<string>Xamarin.Cognitive.Speech.Sample</string>
 	<key>XSAppIconAssets</key>
 	<string>Assets.xcassets/AppIcon.appiconset</string>
 	<key>NSMicrophoneUsageDescription</key>
