--- Xamarin.Cognitive.BingSpeech.NuGet/readme.txt
@@ -1,3 +1,0 @@
-﻿
-View and download/clone the sample app (https://github.com/NateRickard/Xamarin.Cognitive.BingSpeech/tree/master/Sample) 
-for a complete example of how to use this package.
