--- Xamarin.Cognitive.BingSpeech/README.md
@@ -1,2 +1,0 @@
-# Xamarin.Cognitive.BingSpeech
-A client library that makes it easy to work with the Microsoft Cognitive Services Bing Speech API on Xamarin.iOS, Xamarin.Android, and Xamarin.Forms and/or Portable Class Libraries.
