+++ Xamarin.Cognitive.Speech.NuGet/Readme.txt
@@ -1,0 +1,3 @@
+
+View and download/clone the sample app (https://github.com/NateRickard/Xamarin.Cognitive.Speech/tree/master/Sample) 
+for a complete example of how to use this package.
