+++ Xamarin.Cognitive.Speech.NuGet/Xamarin.Cognitive.Speech.NuGet.nuproj
@@ -13,19 +13,19 @@
     </ReferenceAssemblyFrameworks>
   </PropertyGroup>
   <PropertyGroup>
-    <PackageId>Xamarin.Cognitive.BingSpeech</PackageId>
-    <PackageVersion Condition="'$(PackageVersion)' == ''">2.1.3</PackageVersion>
-    <Title>Xamarin client for the Bing Speech API</Title>
+    <PackageId>Xamarin.Cognitive.Speech</PackageId>
+    <PackageVersion Condition="'$(PackageVersion)' == ''">1.0.0</PackageVersion>
+    <Title>Xamarin client for the Azure Cognitive Services Speech Services Speech to Text API</Title>
     <Authors>Nate Rickard</Authors>
     <Owners>Nate Rickard</Owners>
-    <Summary>Xamarin.Cognitive.BingSpeech is a managed client library for the Microsoft Cognitive Services Bing Speech API</Summary>
-    <Description>A managed client library that makes it easy to work with the Microsoft Cognitive Services Bing Speech API on Xamarin.iOS, Xamarin.Android, and Xamarin.Forms, UWP, and/or .NET Standard projects.</Description>
+    <Summary>Xamarin.Cognitive.Speech is a managed client library for the Microsoft Cognitive Services Speech Services Speech to Text API</Summary>
+    <Description>A managed client library that makes it easy to work with the Microsoft Cognitive Services Speech Services Speech to Text API on Xamarin.iOS, Xamarin.Android, and Xamarin.Forms, UWP, and/or .NET Standard projects.</Description>
     <PackageReleaseNotes>
     </PackageReleaseNotes>
-    <PackageProjectUrl>https://github.com/NateRickard/Xamarin.Cognitive.BingSpeech</PackageProjectUrl>
-    <PackageLicenseUrl>https://github.com/NateRickard/Xamarin.Cognitive.BingSpeech/blob/master/LICENSE</PackageLicenseUrl>
+    <PackageProjectUrl>https://github.com/NateRickard/Xamarin.Cognitive.Speech</PackageProjectUrl>
+    <PackageLicenseUrl>https://github.com/NateRickard/Xamarin.Cognitive.Speech/blob/master/LICENSE</PackageLicenseUrl>
     <Copyright>Copyright © Nate Rickard</Copyright>
-    <PackageTags>xamarin, microsoft, android, xamarin.forms, UWP, ios, speech, cognitive, bing, SDK, speech to text, STT, speech recognition, .NET Standard</PackageTags>
+    <PackageTags>xamarin, microsoft, android, xamarin.forms, UWP, ios, speech, speech services, cognitive, bing, SDK, speech to text, STT, speech recognition, .NET Standard</PackageTags>
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
     <TargetFramework>netstandard1.0</TargetFramework>
     <PackageTargetFallback>net11;net20;net35;net40;net403;net45;net451;net452;net46;net461;net462;net47;netcore;netcore45;netcore451;netcore50;win8;win81;win10;sl4;sl5;wp;wp7;wp75;wp8;wp81;wpa81;uap;uap10;netstandard1.0;netstandard1.1;netstandard1.2;netstandard1.3;netstandard1.4;netstandard1.5;netstandard1.6;netstandard2.0;netcoreapp1.0;netcoreapp2.0;monoandroid;monotouch;monomac;xamarinios;xamarinmac;xamarinpsthree;xamarinpsfour;xamarinpsvita;xamarinwatchos;xamarintvos;xamarinxboxthreesixty;xamarinxboxone</PackageTargetFallback>
@@ -44,7 +44,7 @@
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Xamarin.Cognitive.BingSpeech\Xamarin.Cognitive.Speech.csproj" />
+    <ProjectReference Include="..\Xamarin.Cognitive.Speech\Xamarin.Cognitive.Speech.csproj" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.Common.targets" />
   <Import Project="$(NuGetAuthoringPath)\NuGet.Packaging.Authoring.targets" Condition="Exists('$(NuGetAuthoringPath)\NuGet.Packaging.Authoring.targets')" />
