+++ Xamarin.Cognitive.Speech/Api/PushStreamContent.cs
