+++ Xamarin.Cognitive.Speech/Model/OutputMode.cs
@@ -1,4 +1,4 @@
-namespace Xamarin.Cognitive.BingSpeech
+namespace Xamarin.Cognitive.Speech
 {
 	/// <summary>
 	/// Output mode.
