+++ Xamarin.Cognitive.Speech/Model/RecognitionMode.cs
@@ -1,4 +1,4 @@
-namespace Xamarin.Cognitive.BingSpeech
+namespace Xamarin.Cognitive.Speech
 {
 	/// <summary>
 	/// Recognition mode.
