+++ Xamarin.Cognitive.Speech/Model/SpeechResult.cs
@@ -1,4 +1,4 @@
-namespace Xamarin.Cognitive.BingSpeech
+namespace Xamarin.Cognitive.Speech
 {
 	/// <summary>
 	/// A Speech result that is part of a detailed resultset.
