+++ Xamarin.Cognitive.Speech/README.md
@@ -1,0 +1,2 @@
+# Xamarin.Cognitive.Speech
+A client library that makes it easy to work with the Microsoft Cognitive Services Speech Services Speech to Text API on Xamarin.iOS, Xamarin.Android, and Xamarin.Forms and/or .NET Standard Libraries.
