+++ Xamarin.Cognitive.Speech/Values/Endpoints.cs
@@ -1,4 +1,4 @@
-namespace Xamarin.Cognitive.BingSpeech
+namespace Xamarin.Cognitive.Speech
 {
 	static class Endpoints
 	{
