+++ Xamarin.Cognitive.Speech/Values/Keys.cs
@@ -1,4 +1,4 @@
-namespace Xamarin.Cognitive.BingSpeech
+namespace Xamarin.Cognitive.Speech
 {
 	static partial class Constants
 	{
