+++ Xamarin.Cognitive.Speech/Values/MimeTypes.cs
@@ -1,4 +1,4 @@
-namespace Xamarin.Cognitive.BingSpeech
+namespace Xamarin.Cognitive.Speech
 {
 	static partial class Constants
 	{
