+++ Xamarin.Cognitive.Speech/Xamarin.Cognitive.Speech.csproj
